refactor(.): Add initial emitter api and clean up

+++ include/ras/request.h
@@ -14,7 +14,7 @@ struct ras_request_options_s;
  */
 typedef int (ras_request_callback_t)(
   struct ras_request_s *request,
-  unsigned int err,
+  int err,
   void *value,
   unsigned long int size);
 
@@ -41,7 +41,7 @@ enum ras_request_type {
   struct ras_storage_s *storage;    \
   ras_request_callback_t *before;   \
   ras_request_callback_t *after;    \
-  unsigned int offset;              \
+  unsigned long int offset;         \
   unsigned long int size;           \
   void *callback;                   \
   void *data;
@@ -60,8 +60,8 @@ struct ras_request_options_s {
  */
 #define RAS_REQUEST_FIELDS          \
   unsigned int alloc:1;             \
-  unsigned int err;                 \
-  unsigned int offset;              \
+  int err;                          \
+  unsigned long int offset;         \
   unsigned long int size;           \
   unsigned int pending:1;           \
   enum ras_request_type type;       \
@@ -127,7 +127,7 @@ ras_request_run(struct ras_request_s *request);
 RAS_EXPORT int
 ras_request_callback(
   struct ras_request_s *request,
-  unsigned int err,
+  int err,
   void *value,
   unsigned long int size);
